task 6

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -7,7 +7,8 @@
 
 void puts2(char *str)
 {
-	int i = 0;
+	long i = 0;
+
 	for ( ; str[i] != '\0'; i += 2)
 		_putchar(str[i]);
 	_putchar('\n');
